# perf(github): improve `gh` start time by skipping lookup table generation